fix: an unexaminable candidate is a coverage gap, not a filtered file
Codex P2 on c9b18e28, and the third instance of this PR's own bug class:
a silent drop that leaves the account looking complete.

The extensionless-header heuristic has to *read* a file to classify it, and
`_looks_binary` returns True on `OSError` — correct for its own callers, but for
discovery it meant an unreadable extensionless header (an ACL, a transient I/O
error) was classified "binary", dropped by the scannability filter, and never
recorded. A sibling file scanning successfully then let the set report
`sufficient=True` over a header nobody read, and an absent construct could fold
to `introduced`/`resolved`.

Fixed where the conflation lives rather than at the call site: discovery now
classifies **tri-state** — a real candidate, a file that was never evidence (a
`.md`, a binary blob, an oversized extensionless file), or a candidate that
could not be examined. Only the last is a gap; only the middle is silent. A
two-state scannable/not-scannable predicate cannot express that, which is why
`resolve_source_inputs` now takes a classifier returning a disposition or
`None` instead of a bool.

Writing the test found a second, adjacent hole: a candidate whose `is_file()`
itself raises (`is_file` stats, and a stat can fail on its own) propagated the
`OSError` out of an advisory walk. Now recorded as `UNREADABLE` — a walk that
must not abort must also not quietly shrink its own expected set.

Tests parametrised over both ways the heuristic can fail to examine a file (the
stat and the open), injected rather than chmod'd since these suites can run as
root, plus the consequence end to end and the three not-evidence cases that must
*not* become gaps. Verified by restoring the two-state behaviour: three fail.

`tests/test_stored_snapshot_source_licence.py` reached the architecture gate's
test cap, so it is split along the bug-class boundary it already had: the
provenance class stays, `coverage.discovery_derived_completeness` moves to
`tests/test_source_input_completeness.py`, and shared constructors move to
`tests/_source_licence_fixtures.py` (the repo's existing non-collected helper
convention). Split at the boundary, not at an arbitrary line.
